+++ CHANGES.rst
@@ -188,6 +188,9 @@ mast
 - Present users with an error rather than a warning when nonexistent query criteria are used in ``mast.Observations.query_criteria``
   and ``mast.Catalogs.query_criteria``. [#3084]
 
+- Support for case-insensitive criteria keyword arguments in ``mast.Observations.query_criteria`` and 
+  ``mast.Catalogs.query_criteria``. [#3085]
+
 
 0.4.7 (2024-03-08)
 ==================